Create a list of derivatives of expected gpl losses from a gpl_df

Create a list of derivatives of expected gpl losses from a gpl_df

Usage

dexp_gpl_df(df, F)

Arguments

df

a gpl_df, as created by new_gpl_df().

F

list of predictive cdfs, one per row of df.

Value

A list of functions of an allocation x, one per target.

Examples

gdf <- new_gpl_df(N = 2, alpha = 0.5)
dexp_gpl_df(gdf, F = list(pnorm, pnorm))
#> [[1]]
#> function (x) 
#> {
#>     kappa * ((1 - alpha) * dexp_over_loss(dg = dg, F = F)(x) + 
#>         alpha * dexp_under_loss(dg = dg, F = F)(x))
#> }
#> <bytecode: 0x564b94b2dd80>
#> <environment: 0x564b9701ea00>
#> 
#> [[2]]
#> function (x) 
#> {
#>     kappa * ((1 - alpha) * dexp_over_loss(dg = dg, F = F)(x) + 
#>         alpha * dexp_under_loss(dg = dg, F = F)(x))
#> }
#> <bytecode: 0x564b94b2dd80>
#> <environment: 0x564b9701b250>
#>